More tweaks before next release - headless mode (beta), and playlist options
Tweaking the codebase a bit more before the next release (the first new
one in 2 years!) - added new UDF from the AutoIt forums to allow better
and more accurate highlighting of which event is currently looping by
coloring the line the color of the current play mode, and the text in
bold.  Added the ability to play only the playlist items that were
selected (per Web request), and working on a headless (GUI-less) mode
for another request.  That mode *works*, but needs a better integration
into the program so it can be triggered at will... so a few more steps,
some extra fine combing for issues, and a scan... and hopefully will be
able to release this week :)