2-11-19 - Use INI files to check AllowMultipleInstances in MPC-HC
Per #2 issue report, @h0rn3z0r reported that Looper doesn't check the
INI files that MPC-HC may be configured with to see if MPC-HC allows
multiple instances.  Coded in a redundancy check (granted, it checks all
3 paths at the time being whether or not INI files are specified), that
checks both 32-bit and 64-bit paths of MPC-HC for that INI parameter, as
well as the registry entry for the same parameter, and returns > 0 if
any of them have the AllowMultipleInstances flag set to anything other
than 0.