2-20-19 - Stay in current mode even after loading a new list
Added new feature (by request) to stay in the current mode when you're
loading a new file.  TO DO - fix one thing going on when you're in
Randomize mode and there isn't anything there to randomize... most
likely a proviso in doTheLoop to just jump back to Loop mode if you're
just clearing events, or possibly in another area... if you actually
click "Clear Events" instead of using the Load function to do it.  ALSO,
re-randomize the random list in search mode?