2-21-19 - File finding function has saving capabilites now!
Added the askForSave() function in to the file finding function - so now
you can find an unlinked file, synchronize that new path across the
entire event list, and then save the .looper file with the updated
information.  Not sure if I'm 100% happy with how everything works just
yet (Do I pause playback before bringing up the prompt for file finidng?
Do I check the INI setting for pausing playback to check for that?  If
you cancel, or give a weird prompt, do I bow out completely, or put an
error box up?), but the feature does work how it should, and I'm happy
with that.  A little more finessing and it'll be mergable back into
master.