2-22-19 - Opening up the Hood and Tighenin' up the Valves!
Usability tests - running Looper and seeing how things work, how the new
features integrate and fixing little things along the way.  Corrected
Playlist/Shuffle error displays for new features.  Added an option to
the Options pane to not force Loop mode when opening new files
(suggested by @h0rn3z0r in his Issue #2 request).  Added the "repeat n
times" feature to Shuffle mode as well.  Cleaned up some other
eccentricites (not allowing Shuffle mode when only one event in the
events list, keeping the error display on for a longer period of time
when switching to a mode that won't work, lengthening the OSD events
display to properly show both repeats (if there are any) AND the time of
that event, Shuffle Mode not working 100% correctly when in Search mode
as well, etc.) - a bunch of little things here and there.  And now on to
MORE testing before the next release!