2-28-19 - Modularization changes, and a few other things
Trying to keep each function modular, so I broke a few functions out
from scripts that had multiple functions (defaults, switchToLoop,
switchToPlaylist, switchToShuffle, etc...), and also made the requisite
changes in the main .au3 file to reflect the "new" modules.  Also
renamed isAcceptable function to isAcceptableNumber (because it
explicitly deals with numbers) and added an error message to the Options
dialog when you enter something that's not a number into either the
Preview or Slip fields at the top.  Also moved the GUI display on launch
to earlier in the script, to help with sizing windows on launch.