3-7-19 - One more *small* change, getting closer!
One more small change to a behavior noticed while testing - if you don't
have a file, and you choose not to locate it, advancing to the next
event manually won't work, because $currentPlayingEvent doesn't
increment... so it tries to find the same file again (as it just
re-loads that event one more time) - that's been changed.  Getting
closer and closer to release!