3-13-19 - A few after-release tweaks, and a new release with those tweaks...
When I wrote this new version, I added a jump-back to Loop Mode setting
when jumping to a new file in MPC-HC, but this broke (most likely only
used by me) Looper's EDL functions, so I wrote a better conditional for
it.  I also noticed some weird speed setting issues when loading some
files, so I tweaked the load event structuring a bit to better
compensate for that.