4-27-20 - Sidecar file save prompt (by request)
By request for feature request #5

Added check to see if every single event in the events list is from the
same file.  If it is, then when saving the .looper file, it (by default)
chooses the relative path and filename of the media file as its default
location to save the .looper file, making it a "sidecar" file companion
to the media file.  You can of course still save to a different path,
this just changes the *default* behavior when prompting to save.