7-5-20 - More work on MPC-BE multiple instances, removal of refreshMPCInfo(), removed RC .exe, compiled another RC .exe
Tweaked MPC-BE (and PortableApps verison of MPC-HC) identifying whether
or not the *AllowMultipleFlags* (or *MultipleFlags* for MPC-BE)
parameter has been set, allowing multiple Looper/MPC instances.  Also
removed refreshMPCInfo() as MPC-HC/BE _already_ automatically refresh
the current information (it doesn't need to be polled) - only instance
of it remaining is the very beginning of the main loop (so if there _is_
a file playing when you open Looper, it will get the currect IN/OUT
info).