3-18-21 - Fixed nowPlayingInfo requests for files with | in their title/author/etc. information
To correct #7 - Looper now queries the right side instead of the left
side of the nowPlayingInfo array to get the correct media path and
duration (if a media file contained a "|" character in its title,
MPC-HC/BE reports that in the returned information, which changes the
number of elements in the nowPlayingInfo array.  The path name is always
the second to last element, and the duration is the last.