3-19-21 - Fixed Docking procedures
To correct #9 I wrote a procedure to check the location of the taskbar
when the docking feature is first launched, moving it left/right/down by
an offset relative to the taskbar's current location.  Also refined the
initial launch procedure (when you have Looper set to remember Docking
on launch) to check the above procedure to launch in the correct
location.