3-25-21 - Version check / fix for time offsets
For #12 - Added a version check to determine if Looper needs to offset
the time reported from MPC-HC to keep it in line with the time MPC-HC
reports (in non-"High precision" mode).  Older versions of MPC-HC (and
current versions of MPC-BE) report "normal" time as 500ms offset from
what "High precision" mode does, but newer ones report the time
correctly.